requirements.txt
Flask==2.3.1
requests==2.29.0
python-dotenv==1.0.0